:lipstick: AlertModal, TermsModal - 이용약관 동의 모달 - Alert 모달 추상화
