+++ src/components/modals/index.jsx
@@ -1,0 +1,2 @@
+export { default as TermsModal } from './TermsModal';
+export { default as AlertModal } from './AlertModal';
